changed synthetic data filter
